Fixed warnings

+++ Engine/src/Audio.cpp
@@ -61,7 +61,7 @@ void audio_callback(void *userdata, Uint8 *stream, int len) {
 	if (audio_len == 0)
 		return;
 
-	len = (len > audio_len ? audio_len : len);
+	len = (Uint32(len) > audio_len ? audio_len : len);
 	SDL_memcpy(stream, audio_pos, len); 					// simply copy from one buffer into the other
 	//SDL_MixAudio(stream, audio_pos, len, SDL_MIX_MAXVOLUME);// mix from one buffer into another
